web/requirements.txt
django>=4.0.0,<4.2.0
djangorestframework
djangorestframework.simplejwt
pyyaml
requests
django-cors-headers
drf_yasg
django-dotenv
gunicorn